+++ test/noahtheduke/splint/rules/style/single_key_in_test.clj
@@ -4,15 +4,25 @@
 
 (ns noahtheduke.splint.rules.style.single-key-in-test
   (:require
-    [expectations.clojure.test :refer [defexpect expect]]
-    [noahtheduke.splint.test-helpers :refer [check-alt check-message]]))
+    [expectations.clojure.test :refer [defexpect]]
+    [noahtheduke.splint.test-helpers :refer [expect-match]]))
 
 (defexpect single-key-in-test
-  (expect '(assoc coll :k v) (check-alt "(assoc-in coll [:k] v)"))
-  (expect "Use `assoc` instead of recreating it." (check-message "(assoc-in coll [:k] v)"))
-  (expect '(get coll :k) (check-alt "(get-in coll [:k])"))
-  (expect '(get coll :k :default) (check-alt "(get-in coll [:k] :default)"))
-  (expect "Use `get` instead of recreating it." (check-message "(get-in coll [:k])"))
-  (expect '(update coll :k inc) (check-alt "(update-in coll [:k] inc)"))
-  (expect '(update coll :k + 1 2 3) (check-alt "(update-in coll [:k] + 1 2 3)"))
-  (expect "Use `update` instead of recreating it." (check-message "(update-in coll [:k] inc)")))
+  (expect-match
+    '[{:alt (assoc coll :k v)
+       :message "Use `assoc` instead of recreating it." }]
+    "(assoc-in coll [:k] v)")
+  (expect-match
+    '[{:alt (get coll :k)
+       :message "Use `get` instead of recreating it."}]
+    "(get-in coll [:k])")
+  (expect-match
+    '[{:alt (get coll :k :default)}]
+    "(get-in coll [:k] :default)")
+  (expect-match
+    '[{:alt (update coll :k inc)
+       :message "Use `update` instead of recreating it."}]
+    "(update-in coll [:k] inc)")
+  (expect-match
+    '[{:alt (update coll :k + 1 2 3)}]
+    "(update-in coll [:k] + 1 2 3)"))