+++ test/noahtheduke/splint/runner_test.clj
@@ -1,20 +1,18 @@
-(ns noahtheduke.splint.runner-test 
+(ns noahtheduke.splint.runner-test
   (:require
-    [expectations.clojure.test :refer [defexpect expect]]
-    [noahtheduke.splint.test-helpers :refer [check-all]]))
+    [expectations.clojure.test :refer [defexpect]]
+    [noahtheduke.splint.test-helpers :refer [expect-match]]))
 
 (defexpect ignore-rules-test
-  (expect nil?
-    (check-all "#_:splint/disable (+ 1 x)")))
+  (expect-match nil "#_:splint/disable (+ 1 x)"))
 
 (defexpect ignore-genre-test
-  (expect nil?
-    (check-all "#_{:splint/disable [style]} (+ 1 x)")))
+  (expect-match nil "#_{:splint/disable [style]} (+ 1 x)"))
 
 (defexpect ignore-specific-rule-test
-  (expect nil?
-    (check-all "#_{:splint/disable [style/plus-one]} (+ 1 x)")))
+  (expect-match nil "#_{:splint/disable [style/plus-one]} (+ 1 x)"))
 
 (defexpect ignore-unnecessary-rule-test
-  (expect some?
-    (check-all "#_{:splint/disable [style/plus-zero]} (+ 1 x)")))
+  (expect-match
+    '[{:rule-name style/plus-one}]
+    "#_{:splint/disable [style/plus-zero]} (+ 1 x)"))